+++ src/test/validation/unknownIdentifiers.test.ts
@@ -8,7 +8,7 @@ import { expect } from "chai";
 import { Diagnostic, DiagnosticErrorCode, Position, ValidationSettings } from "../../powerquery-language-services";
 import { SimpleValidateNoneSettings, TestLibraryName } from "../testConstants";
 import { TestConstants, TestUtils } from "..";
-import { expectLessWhenSurpressed } from "./common";
+import { expectLessWhenSurpressed } from "../testUtils/validationTestUtils";
 
 interface AbridgedUnknownIdentifierDiagnostic {
     readonly message: string;
